+++ lib/wrapper.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:vizilog/pages/home.dart';
+
+class Wrapper extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Home();
+  }
+}
